iothread: fix poll properties typename
The getter and setter use visit_type_int64, not visit_type_int.

Fixes: 0d9d86fb4df4 ("iothread: add polling parameters")
Message-Id: <20260818-qom-qapi-v3-36-24b8bbbe3d86@redhat.com>

+++ iothread.c
@@ -315,19 +315,19 @@ static void iothread_class_init(ObjectClass *klass, const void *class_data)
     bc->init = iothread_init;
     bc->update_params = iothread_set_aio_context_params;
 
-    object_class_property_add(klass, "poll-max-ns", "int",
+    object_class_property_add(klass, "poll-max-ns", "int64",
                               iothread_get_poll_param,
                               iothread_set_poll_param,
                               NULL, &poll_max_ns_info);
-    object_class_property_add(klass, "poll-grow", "int",
+    object_class_property_add(klass, "poll-grow", "int64",
                               iothread_get_poll_param,
                               iothread_set_poll_param,
                               NULL, &poll_grow_info);
-    object_class_property_add(klass, "poll-shrink", "int",
+    object_class_property_add(klass, "poll-shrink", "int64",
                               iothread_get_poll_param,
                               iothread_set_poll_param,
                               NULL, &poll_shrink_info);
-    object_class_property_add(klass, "poll-weight", "int",
+    object_class_property_add(klass, "poll-weight", "int64",
                               iothread_get_poll_param,
                               iothread_set_poll_param,
                               NULL, &poll_weight_info);
